Refactor
Devide code into smaller components
