+++ package.json
@@ -15,6 +15,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^3.11.0",
     "react-responsive-carousel": "^3.2.23",
+    "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
