+++ src/App.js
@@ -4,17 +4,19 @@ import {
     Box,
     Grid
 } from "@chakra-ui/react";
-import Header from "./components/Header";
 import theme from "./components/theme";
-import Main from "./components/Main";
+
+import Header from "./Layout/Header";
+import Page from "./Layout/Page";
+import { Navigations } from "./Layout/Navigations";
 
 export const App = () => (
     <ChakraProvider theme={theme}>
         <Box textAlign="center" fontSize="xl" background={'#070707'}  w={'376px'} m={'0 auto'}>
             <Grid minH="100vh" pe={6} ps={6} >
+                <Navigations/>
                 <Header/>
-                <Main/>
-        
+                <Page/>
             </Grid>
         </Box>
     </ChakraProvider>
