+++ src/Layout/Page.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { Routes, Route } from "react-router-dom";
+import HomePage from "../pages/HomePage";
+
+const Page = () => {
+    return (
+        <Routes>
+            <Route path="/" element={<HomePage />} />
+        </Routes>
+        
+    );
+};
+
+export default Page;
