+++ src/index.js
@@ -1,12 +1,15 @@
 import React, { StrictMode } from 'react';
 import * as ReactDOM from 'react-dom/client';
 import { App } from './App';
+import { BrowserRouter } from 'react-router-dom';
 
 const container = document.getElementById('root');
 const root = ReactDOM.createRoot(container);
 
 root.render(
     <StrictMode>   
-        <App />
+        <BrowserRouter>
+            <App  />
+        </BrowserRouter>
     </StrictMode>
 );
